Fixing minor things for the Azure backed nb storage.
